# packages/proto/openapi.yaml
openapi: 3.1.0
info:
  title: Exponential Headless API
  version: 0.1.0
...
servers:
  - url: http://localhost:7016/v1
security:
  - bearerAuth: []
paths:
  /sync/status:
    get:
      operationId: getSyncStatus
      summary: Get current workspace sync version and replay status
      responses:
        '200':
          description: Sync status
          content:
            application/json:
              schema:
                type: object
                required: [currentVersion]
                properties:
                  currentVersion:
                    type: integer
                    format: int64
                  lastReplay:
                    type: object
                    required: [requestedVersion, operationCount, replayedAt]
                    properties:
                      requestedVersion:
                        type: integer
                        format: int64
                      operationCount:
                        type: integer
                      replayedAt:
                        type: string
                        format: date-time
        default:
          $ref: '#/components/responses/Problem'
...
  /customers:
    get:
      operationId: listCustomers
      summary: List customers in the authenticated workspace
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: tier
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Customers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomerListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createCustomer
      summary: Create a customer in the authenticated workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomerRequest'
      responses:
        '201':
          description: Customer
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        default:
          $ref: '#/components/responses/Problem'
...
  /demo/session:
    get:
      operationId: createDemoSession
      summary: Create a disposable public demo browser session and seed the demo workspace if needed.
      security: []
      responses:
        '200':
          description: Demo session details for API clients
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DemoSessionResponse'
        '302':
          description: Browser requests are redirected to the seeded demo workspace after cookies are set.
        default:
          $ref: '#/components/responses/Problem'
  /issues:
    get:
      operationId: listIssues
      summary: List issues in the authenticated workspace
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
        - name: team_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Issues page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createIssue
      summary: Create an issue
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueRequest'
      responses:
        '201':
          description: Created issue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Issue'
        default:
          $ref: '#/components/responses/Problem'
...
  /issues/{id}/comments:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: createIssueComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentRequest'
      responses:
        '201':
          description: Created comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        default:
          $ref: '#/components/responses/Problem'
  /issues/{id}/reactions:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    post:
      operationId: toggleIssueReaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionRequest'
      responses:
        '200':
          description: Issue reaction summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionSummaryList'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteIssueReaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionRequest'
      responses:
        '200':
          description: Issue reaction summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionSummaryList'
        default:
          $ref: '#/components/responses/Problem'
...
  /issue-templates:
    get:
      operationId: listIssueTemplates
      parameters:
        - name: teamKey
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Issue templates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueTemplateListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createIssueTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIssueTemplateRequest'
      responses:
        '201':
          description: Created issue template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssueTemplateResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /account/notifications:
    get:
      operationId: getAccountNotifications
      responses:
        '200':
          description: Account notification settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountNotificationsResponse'
        default:
          $ref: '#/components/responses/Problem'
    patch:
      operationId: updateAccountNotifications
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccountNotificationsRequest'
      responses:
        '200':
          description: Updated account notification settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountNotificationsResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /agent/runs:
    get:
      operationId: listAgentRuns
      summary: List workspace agent runs
      responses:
        '200':
          description: Agent runs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createAgentRun
      summary: Create an agent run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRunRequest'
      responses:
        '201':
          description: Created agent run
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /auth/session:
    get:
      operationId: getAuthSession
      summary: Get the authenticated browser session
      responses:
        '200':
          description: Browser session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthSessionResponse'
        default:
          $ref: '#/components/responses/Problem'
  /comments/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateComment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentRequest'
      responses:
        '200':
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteComment
      responses:
        '200':
          description: Deleted comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        default:
          $ref: '#/components/responses/Problem'
  /comments/{id}/reactions:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: toggleCommentReaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReactionRequest'
      responses:
        '200':
          description: Comment reaction summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReactionSummaryList'
        default:
          $ref: '#/components/responses/Problem'
  /custom-emojis:
    get:
      operationId: listCustomEmojis
      responses:
        '200':
          description: Custom emojis
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomEmojiListResponse'
        default:
          $ref: '#/components/responses/Problem'
    post:
      operationId: createCustomEmoji
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCustomEmojiRequest'
      responses:
        '201':
          description: Created custom emoji
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CustomEmojiResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /document-settings:
    get:
      operationId: getDocumentSettings
      responses:
        '200':
          description: Document settings
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSettingsResponse'
        default:
          $ref: '#/components/responses/Problem'
  /document-folders:
    post:
      operationId: createDocumentFolder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentFolderRequest'
      responses:
        '201':
          description: Created document folder
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentFolderResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /document-templates:
    post:
      operationId: createDocumentTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentTemplateRequest'
      responses:
        '201':
          description: Created document template
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentTemplateResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /integrations:
    get:
      operationId: listIntegrations
      responses:
        '200':
          description: Workspace integrations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntegrationListResponse'
        default:
          $ref: '#/components/responses/Problem'
    delete:
      operationId: deleteIntegration
      parameters:
        - name: provider
          in: query
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Deleted integration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /integrations/github/setup/callback:
    get:
      operationId: githubSetupCallback
      security: []
      parameters:
        - name: installation_id
          in: query
          schema: { type: string }
        - name: setup_action
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        '302': { description: Redirects to integration settings }
        default:
          $ref: '#/components/responses/Problem'
  /integrations/github/webhook:
    post:
      operationId: ingestGitHubWebhook
      security: []
      parameters:
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema: { type: string }
        - name: X-GitHub-Delivery
          in: header
          required: true
          schema: { type: string }
        - name: X-GitHub-Event
          in: header
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '202':
          description: GitHub webhook accepted or deterministically ignored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitHubWebhookResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /integrations/gitlab/webhook/{integrationId}:
    post:
      operationId: ingestGitLabWebhook
      security: []
      parameters:
        - name: integrationId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '202':
          description: GitLab webhook accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitLabWebhookResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /integrations/google-sheets/oauth/callback:
    get:
      operationId: googleSheetsOAuthCallback
      security: []
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        '302': { description: Redirects to integration settings }
        default:
          $ref: '#/components/responses/Problem'
...
  /integrations/intercom/oauth/callback:
    get:
      operationId: intercomOAuthCallback
      security: []
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        '302': { description: Redirects to integration settings }
        default:
          $ref: '#/components/responses/Problem'
  /integrations/intercom/issues/search:
    post:
      operationId: searchIntercomLinkableIssues
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntercomIssueActionRequest'
      responses:
        '200':
          description: Linkable Exponential issues for Intercom
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntercomIssueSearchResponse'
        default:
          $ref: '#/components/responses/Problem'
  /integrations/intercom/issues/status:
    post:
      operationId: getIntercomLinkedIssueStatus
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntercomIssueActionRequest'
      responses:
        '200':
          description: Linked issue status for an Intercom conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntercomIssueActionResponse'
        default:
          $ref: '#/components/responses/Problem'
  /integrations/intercom/issues/link:
    post:
      operationId: linkIntercomIssue
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntercomIssueActionRequest'
      responses:
        '200':
          description: Linked issue for an Intercom conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntercomIssueActionResponse'
        default:
          $ref: '#/components/responses/Problem'
  /integrations/intercom/issues/unlink:
    post:
      operationId: unlinkIntercomIssue
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntercomIssueActionRequest'
      responses:
        '200':
          description: Unlinked issue for an Intercom conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntercomIssueActionResponse'
        default:
          $ref: '#/components/responses/Problem'
  /integrations/intercom/issues/create:
    post:
      operationId: createIssueFromIntercom
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IntercomIssueActionRequest'
      responses:
        '200':
          description: Created issue from an Intercom conversation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IntercomIssueActionResponse'
        default:
          $ref: '#/components/responses/Problem'
...
  /integrations/sentry/oauth/callback:
    get:
      operationId: sentryOAuthCallback
      security: []
      parameters:
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          schema: { type: string }
      responses:
        '302': { description: Redirects to integration settings }
        default:
          $ref: '#/components/responses/Problem'
...
  /test:
    get:
      operationId: testHelpersMount
      responses:
        '404':
          description: Test helper mount marker
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: lin_api or pat
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string, minLength: 1, maxLength: 255 }
  responses:
    Problem:
      description: RFC 7807 problem response
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    CreateCommentRequest:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1 }
...
    CustomerListResponse:
      type: object
      required: [customers]
      properties:
        customers:
          type: array
          items:
            $ref: '#/components/schemas/Customer'
...
    CreateCustomerRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        domain: { type: [string, 'null'] }
        revenue: { type: [number, 'null'] }
        size: { type: [integer, 'null'] }
        tier: { type: [string, 'null'] }
        status: { type: [string, 'null'] }
        ownerId: { type: [string, 'null'] }
        source: { type: [string, 'null'] }
...
    IssueListResponse:
      type: object
      required: [data]
      properties:
        data:
          type: array
          items: { $ref: '#/components/schemas/Issue' }
        next_cursor: { type: [string, 'null'] }
...